fastapi_be/requirements.txt
uvicorn
fastapi
email-validator
mcp
openai
python-dotenv